requirements.txt
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Pillow==10.0.1
pydantic==2.3.0
pydantic_core==2.6.3
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
requests==2.31.0
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.8.0
urllib3==2.0.5
